requirements.txt
streamlit==1.32.0
python-dotenv==1.0.0
langchain==0.1.12
openai==1.14.0
prefect==2.14.21
fastapi==0.110.0
uvicorn==0.27.1
sqlalchemy==2.0.28
slack-sdk==3.27.1
github3.py==4.0.1
pandas==2.2.1
pyarrow==15.0.0
boto3==1.34.69
requests==2.31.0
python-jose==3.3.0
pydantic==2.6.3
jira==3.5.2 